pyproject.toml
[project]
name = "transcribe-diarize"
version = "0.1.0"
description = "Transcribe audio files with speaker labels using Whisper and pyannote.audio"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "openai-whisper",
    "pyannote.audio==4.0.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "python-dotenv",
]

[project.scripts]
transcribe-diarize = "transcribe_diarize:main"
